+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+ * Add support for the Manjaro kernel
+   ([#165](https://github.com/rnestler/reboot-arch-btw/pull/165))
+
 ## [v0.6.0] - 2024-01-20
 
  * Print the cleaned kernel version